exercicios/10_triagem_imagens_patologia_sem_ecra/requirements.txt
python-dotenv>=1.0.0,<2
langgraph>=0.2.0,<2
langchain-core>=0.3.0
langchain-google-genai>=4.0.0,<5
pymongo>=4.6.0,<5
numpy>=1.26.0,<3
scikit-learn>=1.3.0,<2
joblib>=1.3.0,<2
medmnist>=3.0.0,<4
torch>=2.2.0,<2.10
torchvision>=0.17.0,<1
Pillow>=10.0.0,<12
